+++ addresses/avax/elk.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x9E4AAbd2B3E60Ee1322E94307d0776F2c8e6CFbb",
+  "factory": "0x091d35d7F63487909C863001ddCA481c6De47091"
+}
