+++ addresses/avax/joe.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x60aE616a2155Ee3d9A68541Ba4544862310933d4",
+  "factory": "0x9Ad6C38BE94206cA50bb0d90783181662f0Cfa10"
+}
