+++ addresses/avax/unifi.json
@@ -1,0 +1,4 @@
+{
+  "router": "0xBd562d5cF2c62Da3143D862aF39eDb6dF59A4679",
+  "factory": "0x839547067bc885db205F5fA42dcFeEcDFf5A8530"
+}
