+++ addresses/bsc/jet.json
@@ -1,0 +1,4 @@
+{
+  "router": "0xBe65b8f75B9F20f4C522e0067a3887FADa714800",
+  "factory": "0x0eb58E5c8aA63314ff5547289185cC4583DfCBD5"
+}
