+++ addresses/bsc/kyber.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x78df70615ffc8066cc0887917f2Cd72092C86409",
+  "factory": "0x878dFE971d44e9122048308301F540910Bbd934c"
+}
