+++ addresses/bsc/unifi.json
@@ -1,0 +1,4 @@
+{
+  "router": "0xBE930734eDAfc41676A76d2240f206Ed36dafbA2",
+  "factory": "0xA5Ba037Ec16c45f8ae09e013C1849554C01385f5"
+}
