+++ addresses/fantom/dfyn.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x2724B9497b2cF3325C6BE3ea430b3cec34B5Ef2d",
+  "factory": "0xd9820a17053d6314B20642E465a84Bf01a3D64f5"
+}
