+++ addresses/fantom/elk.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x4D2cf285a519261F30b4d9c2c344Baf260d65Fa2",
+  "factory": "0x7Ba73c99e6f01a37f3e33854c8F544BbbadD3420"
+}
