+++ addresses/fantom/empire.json
@@ -1,0 +1,4 @@
+{
+  "router": "0xdADaae6cDFE4FA3c35d54811087b3bC3Cd60F348",
+  "factory": "0x06530550A48F990360DFD642d2132354A144F31d"
+}
