+++ addresses/fantom/hyper.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x53c153a0df7E050BbEFbb70eE9632061f12795fB",
+  "factory": "0x991152411A7B5A14A8CF0cDDE8439435328070dF"
+}
