+++ addresses/fantom/jet.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x845E76A8691423fbc4ECb8Dd77556Cb61c09eE25",
+  "factory": "0xf6488205957f0b4497053d6422F49e27944eE3Dd"
+}
