+++ addresses/fantom/kyber.json
@@ -1,0 +1,4 @@
+{
+  "router": "0x5d5A5a0a465129848c2549669e12cDC2f8DE039A",
+  "factory": "0x78df70615ffc8066cc0887917f2Cd72092C86409"
+}
