--- addresses/pancake-testnet.json
@@ -1,4 +1,0 @@
-{
-  "router": "0xD99D1c33F9fC3444f8101754aBC46c52416550D1",
-  "factory": "0x6725F303b657a9451d8BA641348b6761A6CC7a17"
-}
