+++ addresses/polygon/dfyn.json
@@ -1,0 +1,4 @@
+{
+  "router": "0xA102072A4C07F06EC3B4900FDC4C7B80b6c57429",
+  "factory": "0xE7Fb3e833eFE5F9c441105EB65Ef8b261266423B"
+}
