+++ addresses/polygon/empire.json
@@ -1,0 +1,4 @@
+{
+  "router": "0xB2855A6dAeeBDB72B0176A479A983066ae9775A6",
+  "factory": "0x06530550A48F990360DFD642d2132354A144F31d"
+}
